+++ src/despeckle.c
@@ -578,7 +578,7 @@ static void despeckle_iteration( /* in */ int level,
   current_size = 1 << level;
   tightness = (int)(noise_max / (1.0 + adaptive_tightness * level));
 
-  mask = g_malloc0(width * height * sizeof(unsigned char));
+  mask = g_malloc0((gsize)width * height * sizeof(unsigned char));
   for (y = 0; y < height; y++) {
     for (x = 0; x < width; x++) {
       if (mask[y * width + x] == 0) {
@@ -627,7 +627,7 @@ static void despeckle_iteration_8( /* in */ int level,
   current_size = 1 << level;
   tightness = (int)(noise_max / (1.0 + adaptive_tightness * level));
 
-  mask = g_malloc0(width * height * sizeof(unsigned char));
+  mask = g_malloc0((gsize)width * height * sizeof(unsigned char));
   for (y = 0; y < height; y++) {
     for (x = 0; x < width; x++) {
       if (mask[y * width + x] == 0) {
